feat(chaos): extract csm-chaos crate and port to no_std
- Extract Slhm2d and ChaoticLsh to standalone csm-chaos crate
- Implement no_std support using libm and alloc
- Maintain backward compatibility in csm-core via feature-gated re-exports
- Port benchmarks and unit tests
