+++ crates/csm-chaos/src/lib.rs
@@ -1,0 +1,9 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
+extern crate alloc;
+
+pub mod hashing;
+pub mod maps;
+
+pub use hashing::chaotic_lsh::ChaoticLsh;
+pub use maps::hyperchaotic::Slhm2d;
