+++ crates/csm-chaos/src/maps/mod.rs
@@ -1,0 +1,1 @@
+pub mod hyperchaotic;
